+++ command/sub8_launch/launch/subsystems/perception.launch
@@ -7,8 +7,9 @@
   <!-- <rosparam file="$(find sub8_launch)/config/default_color_calibration.yaml" /> -->
   <rosparam file="$(find sub8_launch)/config/$(arg color_calibration)" />
 
-  <node pkg="sub8_perception" type="pcl_perception" name="sub_perception" output="screen" />
   <node pkg="sub8_perception" type="follow_orange_pipes.py" name="sub_orange_pipes" output="screen" />
   <node pkg="sub8_perception" type="buoy_2d.py" name="buoy_2d" output="screen" />
+  <node pkg="sub8_perception" type="torpedo_board_pose_est.py" name="torpedo_pose_est" output="screen" />
+  <node pkg="sub8_perception" type="torpedos_cpp" name="torpedos_cc" output="screen" />
 
 </launch>
