+++ gnc/sub8_perception/ros_tools/image_dumper.py
