+++ gnc/sub8_perception/src/sub8_vision_lib/cv_param_helpers.cc
@@ -1,4 +1,4 @@
-#include <sub8_pcl/cv_tools.hpp>
+#include <sub8_vision_lib/cv_tools.hpp>
 #include <ros/console.h>
 
 namespace sub {
