+++ gnc/sub8_perception/src/sub8_vision_lib/visualization.cc
@@ -1,4 +1,4 @@
-#include <sub8_pcl/visualization.hpp>
+#include <sub8_vision_lib/visualization.hpp>
 #include <visualization_msgs/Marker.h>
 
 namespace sub {
