+++ simulation/sub8_gazebo/launch/duck.launch
@@ -95,4 +95,6 @@
   <group ns="/down/left" if="$(arg cameras)">
     <node pkg="image_proc" type="image_proc" name="down_image_proc" />
   </group>
+  
+  <include file="$(find sub8_launch)/launch/subsystems/perception.launch" />
 </launch>
